added CSS and fixed errors
