+++ src/App.tsx
@@ -6,12 +6,10 @@ function App() {
 
   return (
     <>
-    <>
      <Header />
       <main>
         <Outlet />
       </main>
-    </>
     </>
   )
 }
