+++ src/pages/Contact.tsx
@@ -1,0 +1,6 @@
+
+export default function Contact() {
+  return (
+    <div>Contact</div>
+  )
+}
