+++ src/pages/Error.tsx
@@ -1,11 +1,11 @@
 import { useRouteError } from "react-router-dom";
 
 export default function ErrorPage() {
-    const error = useRouteError();
+    const error: any = useRouteError();
     console.error(error);
 
     return (
-        <div id="errror-page">
+        <div id="error-page">
             
             <h1>Oops!</h1>
